refactored code base to only support one type of entity IDs. previously we supported dual types. integer and Long. now we only support Long entity Ids. this is mainly for code organization and readability.

+++ um-webapp/src/main/java/gov/nara/um/persistence/dao/ITaskDAO.java
@@ -5,7 +5,7 @@ import gov.nara.um.persistence.model.Task;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface ITaskDAO extends JpaRepository<Task, Integer>, JpaSpecificationExecutor<Task>, IByNameApi<Task> {
+public interface ITaskDAO extends JpaRepository<Task, Long>, JpaSpecificationExecutor<Task>, IByNameApi<Task> {
             public Task findByName(String name);
 
 }
